+++ libsolidity/analysis/TypeChecker.cpp
@@ -358,8 +358,13 @@ bool TypeChecker::visit(FunctionDefinition const& _function)
 		}
 		else
 		{
-			if (!type(var)->canLiveOutsideStorage() && _function.isPublic())
-				m_errorReporter.typeError(var.location(), "Type is required to live outside storage.");
+			if (
+				!type(var)->canLiveOutsideStorage() &&
+				_function.isPublic() &&
+				!isLibraryFunction
+				)
+					m_errorReporter.typeError(var.location(), "Type is required to live outside storage.");
+
 			if (_function.isPublic())
 			{
 				auto iType = type(var)->interfaceType(isLibraryFunction);